Added threading and exception handling for thread
Added threading and exception handling for the non-gui thread. Also fixed main (I think), and integrated the new class-based structure with the GUI. If you want the GUI to run the test GUI method, pass True as a second parameter to the Root_Win. Still need to add progress bar/message and error logs